dom/tests/mochitest/ajax/offline/test_noManifest.html

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>bad manifest content type</title>

<script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script type="text/javascript">

function expectInvalidState(fn, desc) {
  var gotInvalidState = false;
  try {
    fn();
  } catch(e) {
    if (e.code == DOMException.INVALID_STATE_ERR) {
      gotInvalidState = true;
    }
  }

  ok(gotInvalidState, desc);
}

is(typeof(applicationCache), "object");
is(applicationCache.mozLength, 0, "applicationCache.mozLength should be 0");
is(applicationCache.status, 0, "applicationCache.status should be 0 (UNCACHED)");

expectInvalidState(function() { applicationCache.update(); },
                   "applicationCache.update should throw INVALID_STATE_ERR");
expectInvalidState(function() { applicationCache.swapCache(); },
                   "applicationCache.update should throw INVALID_STATE_ERR");


</script>

</head>

<body>

</body>
</html>
